[#350] added abstract regressor class
